refactoring for multiple chain support
