+++ Dockerfile-stub
@@ -3,5 +3,5 @@
 FROM debian:stable-slim
 WORKDIR /app
 COPY bin/* /app
-ENTRYPOINT ["./eth-beacon-genesis"]
+ENTRYPOINT ["./eth-genesis-state-generator"]
 CMD []
