+++ beaconchain/electra.go
@@ -1,4 +1,4 @@
-package generator
+package beaconchain
 
 import (
 	"fmt"
@@ -7,6 +7,8 @@ import (
 	"github.com/attestantio/go-eth2-client/spec"
 	"github.com/attestantio/go-eth2-client/spec/altair"
 	"github.com/attestantio/go-eth2-client/spec/bellatrix"
+	"github.com/attestantio/go-eth2-client/spec/deneb"
+	"github.com/attestantio/go-eth2-client/spec/electra"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -19,7 +21,7 @@ import (
 	dynssz "github.com/pk910/dynamic-ssz"
 )
 
-type bellatrixBuilder struct {
+type electraBuilder struct {
 	elGenesis       *core.Genesis
 	clConfig        *config.Config
 	dynSsz          *dynssz.DynSsz
@@ -27,23 +29,23 @@ type bellatrixBuilder struct {
 	validators      []*validators.Validator
 }
 
-func NewBellatrixBuilder(elGenesis *core.Genesis, clConfig *config.Config) GenesisBuilder {
-	return &bellatrixBuilder{
+func NewElectraBuilder(elGenesis *core.Genesis, clConfig *config.Config) BeaconGenesisBuilder {
+	return &electraBuilder{
 		elGenesis: elGenesis,
 		clConfig:  clConfig,
 		dynSsz:    beaconutils.GetDynSSZ(clConfig),
 	}
 }
 
-func (b *bellatrixBuilder) SetShadowForkBlock(block *types.Block) {
+func (b *electraBuilder) SetShadowForkBlock(block *types.Block) {
 	b.shadowForkBlock = block
 }
 
-func (b *bellatrixBuilder) AddValidators(val []*validators.Validator) {
+func (b *electraBuilder) AddValidators(val []*validators.Validator) {
 	b.validators = append(b.validators, val...)
 }
 
-func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
+func (b *electraBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 	genesisBlock := b.shadowForkBlock
 	if genesisBlock == nil {
 		genesisBlock = b.elGenesis.ToBlock()
@@ -58,17 +60,31 @@ func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 
 	baseFee, _ := uint256.FromBig(genesisBlock.BaseFee())
 
+	var withdrawalsRoot phase0.Root
+
+	if genesisBlock.Withdrawals() != nil {
+		root, err := beaconutils.ComputeWithdrawalsRoot(genesisBlock.Withdrawals(), b.clConfig)
+		if err != nil {
+			return nil, fmt.Errorf("failed to compute withdrawals root: %w", err)
+		}
+
+		withdrawalsRoot = root
+	}
+
 	transactionsRoot, err := beaconutils.ComputeTransactionsRoot(genesisBlock.Transactions(), b.clConfig)
 	if err != nil {
 		return nil, fmt.Errorf("failed to compute transactions root: %w", err)
 	}
 
-	baseFeeBytes := baseFee.Bytes32()
-	for i, j := 0, len(baseFeeBytes)-1; i < j; i, j = i+1, j-1 {
-		baseFeeBytes[i], baseFeeBytes[j] = baseFeeBytes[j], baseFeeBytes[i]
+	if genesisBlock.BlobGasUsed() == nil {
+		return nil, fmt.Errorf("execution-layer Block has missing blob-gas-used field")
 	}
 
-	execHeader := &bellatrix.ExecutionPayloadHeader{
+	if genesisBlock.ExcessBlobGas() == nil {
+		return nil, fmt.Errorf("execution-layer Block has missing excess-blob-gas field")
+	}
+
+	execHeader := &deneb.ExecutionPayloadHeader{
 		ParentHash:       phase0.Hash32(genesisBlock.ParentHash()),
 		FeeRecipient:     bellatrix.ExecutionAddress(genesisBlock.Coinbase()),
 		StateRoot:        phase0.Root(genesisBlock.Root()),
@@ -79,9 +95,12 @@ func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		GasUsed:          genesisBlock.GasUsed(),
 		Timestamp:        genesisBlock.Time(),
 		ExtraData:        extra,
-		BaseFeePerGas:    baseFeeBytes,
+		BaseFeePerGas:    baseFee,
 		BlockHash:        phase0.Hash32(genesisBlockHash),
 		TransactionsRoot: transactionsRoot,
+		WithdrawalsRoot:  withdrawalsRoot,
+		BlobGasUsed:      *genesisBlock.BlobGasUsed(),
+		ExcessBlobGas:    *genesisBlock.ExcessBlobGas(),
 	}
 
 	depositRoot, err := beaconutils.ComputeDepositRoot(b.clConfig)
@@ -96,14 +115,17 @@ func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		syncCommitteeMaskBytes++
 	}
 
-	genesisBlockBody := &bellatrix.BeaconBlockBody{
+	genesisBlockBody := &electra.BeaconBlockBody{
 		ETH1Data: &phase0.ETH1Data{
 			BlockHash: make([]byte, 32),
 		},
 		SyncAggregate: &altair.SyncAggregate{
 			SyncCommitteeBits: make([]byte, syncCommitteeMaskBytes),
 		},
-		ExecutionPayload: &bellatrix.ExecutionPayload{},
+		ExecutionPayload: &deneb.ExecutionPayload{
+			BaseFeePerGas: uint256.NewInt(0),
+		},
+		ExecutionRequests: &electra.ExecutionRequests{},
 	}
 
 	genesisBlockBodyRoot, err := b.dynSsz.HashTreeRoot(genesisBlockBody)
@@ -127,10 +149,10 @@ func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		minGenesisTime = genesisBlock.Time()
 	}
 
-	genesisState := &bellatrix.BeaconState{
+	genesisState := &electra.BeaconState{
 		GenesisTime:           minGenesisTime + genesisDelay,
 		GenesisValidatorsRoot: validatorsRoot,
-		Fork:                  GetStateForkConfig(spec.DataVersionBellatrix, b.clConfig),
+		Fork:                  GetStateForkConfig(spec.DataVersionElectra, b.clConfig),
 		LatestBlockHeader: &phase0.BeaconBlockHeader{
 			BodyRoot: genesisBlockBodyRoot,
 		},
@@ -157,27 +179,27 @@ func (b *bellatrixBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 	}
 
 	versionedState := &spec.VersionedBeaconState{
-		Version:   spec.DataVersionBellatrix,
-		Bellatrix: genesisState,
+		Version: spec.DataVersionElectra,
+		Electra: genesisState,
 	}
 
-	logrus.Infof("genesis version: bellatrix")
+	logrus.Infof("genesis version: electra")
 	logrus.Infof("genesis time: %v", genesisState.GenesisTime)
 	logrus.Infof("genesis validators root: 0x%x", genesisState.GenesisValidatorsRoot)
 
 	return versionedState, nil
 }
 
-func (b *bellatrixBuilder) Serialize(state *spec.VersionedBeaconState, contentType http.ContentType) ([]byte, error) {
-	if state.Version != spec.DataVersionBellatrix {
+func (b *electraBuilder) Serialize(state *spec.VersionedBeaconState, contentType http.ContentType) ([]byte, error) {
+	if state.Version != spec.DataVersionElectra {
 		return nil, fmt.Errorf("unsupported version: %s", state.Version)
 	}
 
 	switch contentType {
 	case http.ContentTypeSSZ:
-		return b.dynSsz.MarshalSSZ(state.Bellatrix)
+		return b.dynSsz.MarshalSSZ(state.Electra)
 	case http.ContentTypeJSON:
-		return state.Bellatrix.MarshalJSON()
+		return state.Electra.MarshalJSON()
 	default:
 		return nil, fmt.Errorf("unsupported content type: %s", contentType)
 	}
