+++ beaconchain/phase0.go
@@ -1,23 +1,22 @@
-package generator
+package beaconchain
 
 import (
 	"fmt"
 
 	"github.com/attestantio/go-eth2-client/http"
 	"github.com/attestantio/go-eth2-client/spec"
-	"github.com/attestantio/go-eth2-client/spec/altair"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
-	dynssz "github.com/pk910/dynamic-ssz"
 	"github.com/sirupsen/logrus"
 
 	"github.com/ethpandaops/eth-beacon-genesis/beaconutils"
 	"github.com/ethpandaops/eth-beacon-genesis/config"
 	"github.com/ethpandaops/eth-beacon-genesis/validators"
+	dynssz "github.com/pk910/dynamic-ssz"
 )
 
-type altairBuilder struct {
+type phase0Builder struct {
 	elGenesis       *core.Genesis
 	clConfig        *config.Config
 	dynSsz          *dynssz.DynSsz
@@ -25,23 +24,23 @@ type altairBuilder struct {
 	validators      []*validators.Validator
 }
 
-func NewAltairBuilder(elGenesis *core.Genesis, clConfig *config.Config) GenesisBuilder {
-	return &altairBuilder{
+func NewPhase0Builder(elGenesis *core.Genesis, clConfig *config.Config) BeaconGenesisBuilder {
+	return &phase0Builder{
 		elGenesis: elGenesis,
 		clConfig:  clConfig,
 		dynSsz:    beaconutils.GetDynSSZ(clConfig),
 	}
 }
 
-func (b *altairBuilder) SetShadowForkBlock(block *types.Block) {
+func (b *phase0Builder) SetShadowForkBlock(block *types.Block) {
 	b.shadowForkBlock = block
 }
 
-func (b *altairBuilder) AddValidators(val []*validators.Validator) {
+func (b *phase0Builder) AddValidators(val []*validators.Validator) {
 	b.validators = append(b.validators, val...)
 }
 
-func (b *altairBuilder) BuildState() (*spec.VersionedBeaconState, error) {
+func (b *phase0Builder) BuildState() (*spec.VersionedBeaconState, error) {
 	genesisBlock := b.shadowForkBlock
 	if genesisBlock == nil {
 		genesisBlock = b.elGenesis.ToBlock()
@@ -59,19 +58,9 @@ func (b *altairBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		return nil, fmt.Errorf("failed to compute deposit root: %w", err)
 	}
 
-	syncCommitteeSize := b.clConfig.GetUintDefault("SYNC_COMMITTEE_SIZE", 512)
-	syncCommitteeMaskBytes := syncCommitteeSize / 8
-
-	if syncCommitteeSize%8 != 0 {
-		syncCommitteeMaskBytes++
-	}
-
-	genesisBlockBody := &altair.BeaconBlockBody{
+	genesisBlockBody := &phase0.BeaconBlockBody{
 		ETH1Data: &phase0.ETH1Data{
 			BlockHash: make([]byte, 32),
-		},
-		SyncAggregate: &altair.SyncAggregate{
-			SyncCommitteeBits: make([]byte, syncCommitteeMaskBytes),
 		},
 	}
 
@@ -82,11 +71,6 @@ func (b *altairBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 
 	clValidators, validatorsRoot := beaconutils.GetGenesisValidators(b.clConfig, b.validators)
 
-	syncCommittee, err := beaconutils.GetGenesisSyncCommittee(b.clConfig, clValidators, phase0.Hash32(genesisBlockHash))
-	if err != nil {
-		return nil, fmt.Errorf("failed to get genesis sync committee: %w", err)
-	}
-
 	genesisDelay := b.clConfig.GetUintDefault("GENESIS_DELAY", 604800)
 	blocksPerHistoricalRoot := b.clConfig.GetUintDefault("SLOTS_PER_HISTORICAL_ROOT", 8192)
 	epochsPerSlashingVector := b.clConfig.GetUintDefault("EPOCHS_PER_SLASHINGS_VECTOR", 8192)
@@ -96,10 +80,10 @@ func (b *altairBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		minGenesisTime = genesisBlock.Time()
 	}
 
-	genesisState := &altair.BeaconState{
+	genesisState := &phase0.BeaconState{
 		GenesisTime:           minGenesisTime + genesisDelay,
 		GenesisValidatorsRoot: validatorsRoot,
-		Fork:                  GetStateForkConfig(spec.DataVersionAltair, b.clConfig),
+		Fork:                  GetStateForkConfig(spec.DataVersionPhase0, b.clConfig),
 		LatestBlockHeader: &phase0.BeaconBlockHeader{
 			BodyRoot: genesisBlockBodyRoot,
 		},
@@ -117,35 +101,30 @@ func (b *altairBuilder) BuildState() (*spec.VersionedBeaconState, error) {
 		Validators:                  clValidators,
 		Balances:                    beaconutils.GetGenesisBalances(b.clConfig, b.validators),
 		Slashings:                   make([]phase0.Gwei, epochsPerSlashingVector),
-		PreviousEpochParticipation:  make([]altair.ParticipationFlags, len(clValidators)),
-		CurrentEpochParticipation:   make([]altair.ParticipationFlags, len(clValidators)),
-		InactivityScores:            make([]uint64, len(clValidators)),
-		CurrentSyncCommittee:        syncCommittee,
-		NextSyncCommittee:           syncCommittee,
 	}
 
 	versionedState := &spec.VersionedBeaconState{
-		Version: spec.DataVersionAltair,
-		Altair:  genesisState,
+		Version: spec.DataVersionPhase0,
+		Phase0:  genesisState,
 	}
 
-	logrus.Infof("genesis version: altair")
+	logrus.Infof("genesis version: phase0")
 	logrus.Infof("genesis time: %v", genesisState.GenesisTime)
 	logrus.Infof("genesis validators root: 0x%x", genesisState.GenesisValidatorsRoot)
 
 	return versionedState, nil
 }
 
-func (b *altairBuilder) Serialize(state *spec.VersionedBeaconState, contentType http.ContentType) ([]byte, error) {
-	if state.Version != spec.DataVersionAltair {
+func (b *phase0Builder) Serialize(state *spec.VersionedBeaconState, contentType http.ContentType) ([]byte, error) {
+	if state.Version != spec.DataVersionPhase0 {
 		return nil, fmt.Errorf("unsupported version: %s", state.Version)
 	}
 
 	switch contentType {
 	case http.ContentTypeSSZ:
-		return b.dynSsz.MarshalSSZ(state.Altair)
+		return b.dynSsz.MarshalSSZ(state.Phase0)
 	case http.ContentTypeJSON:
-		return state.Altair.MarshalJSON()
+		return state.Phase0.MarshalJSON()
 	default:
 		return nil, fmt.Errorf("unsupported content type: %s", contentType)
 	}
